frontend/requirements.txt
azure-cognitiveservices-speech==1.37.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
injector==0.21.0
numpy==1.26.4
pvporcupine==3.0.2
PyAudio==0.2.14
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
webrtcvad==2.0.10
